fix! product stock conversion and new product category assignment
The stock property in `AddProductsUiMappers.kt` no longer filters characters before converting to `Double`.
In `ProductsDomainToDataMappers.kt`, new products now correctly set `categoryId` to `null` instead of attempting to use `this.category?.id`.
The `stock` property in `UpdateProductUiMappers.kt` now directly uses the `Double` value instead of converting from a `String`.
A default value for `detailsText` was changed in `AddProductComposables.kt`.

+++ demosales-ui/products/src/main/java/com/imecatro/products/ui/update/mappers/UpdateProductUiMappers.kt
@@ -15,7 +15,7 @@ fun UpdateProductUiModel.toDomain(): ProductDomainModel {
         currency = this.currency,
         unit = this.unit,
         details = this.details,
-        stock = ProductStockDomainModel(quantity = this.stock.toDouble(), cost = 0.0, emptyList()),
+        stock = ProductStockDomainModel(quantity = this.stock, cost = 0.0, emptyList()),
         imageUri = this.imageUri?.toString(),
         category = this.category?.let { ProductCategoryDomainModel(name = it) }
     )
